# src/pkg/server/handler/apihandler/embedded/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SMS Service API",
    "version": "1.0.0",
    "license": {
      "name": "unlicensed",
      "url": "www.example.com"
    },
    "contact": {
      "name": "API Support",
      "email": "support@example.com",
      "url": "http://example.com"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "paths": {
...
    "/email/basic/send": {
      "post": {
        "description": "Send an Email with or without attachments by giving the recipients data as well as the message content",
        "summary": "Send an Email with custom text",
        "operationId": "sendEmail",
        "tags": [
          "Email"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email sent"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": []
      }
    },
    "/email/templates/{templateName}": {
      "get": {
        "description": "Get a specific template by its name",
        "summary": "Get Template by Name",
        "operationId": "getTemplateByName",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailTemplate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find template (template name)"
          }
        },
        "security": []
      },
      "post": {
        "description": "Add a new template to the db",
        "summary": "Add new template",
        "operationId": "addNewTemplate",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to add",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added template (name)"
          },
          "400": {
            "description": "Invalid name supplied"
          }
        },
        "security": []
      }
    },
    "/email/templates/{templateName}/send": {
      "post": {
        "description": "Send a templated Email with or without by giving the recipients data as well as the placeholder values",
        "summary": "Send a templated Email",
        "operationId": "sendTemplatedEmail",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to add",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplateFillSendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email sent"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": []
      }
    },
    "/email/templates/{templateName}/placeholders": {
      "get": {
        "description": "Get all placeholders of a specific template that have to be filled with some data",
        "summary": "Get Template Placeholders",
        "operationId": "getTemplatePlaceholdersByName",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPlaceholders"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find template (template name)"
          }
        },
        "security": []
      }
    },
    "/email/templates/{templateName}/fill": {
      "post": {
        "description": "Fill the placeholders of a specific template.",
        "summary": "Fill placeholders of template",
        "operationId": "fillTemplate",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to fill",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FilledEmailTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name or data supplied"
          }
        },
        "security": []
      }
    },
...
    "/pdf/templates/{templateName}": {
      "get": {
        "description": "Get a specific PDF template by its name",
        "summary": "Get PDF Template by Name",
        "operationId": "getPDFTemplateByName",
        "tags": [
          "PDF"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PDFTemplate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find template (template name)"
          }
        },
        "security": []
      },
      "post": {
        "description": "Add a new PDF template to the db. If you want new lines in your template, use '\\n' in the string. For placeholders use the syntax '{{.placeholderName}}'.",
        "summary": "Add new PDF template",
        "operationId": "addNewPDFTemplate",
        "tags": [
          "PDF"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to add",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PDFTemplatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added template (name)"
          },
          "400": {
            "description": "Invalid name supplied"
          }
        },
        "security": []
      }
    },
    "/pdf/templates/{templateName}/placeholders": {
      "get": {
        "description": "Get all placeholders of a specific PDF template that have to be filled with some data",
        "summary": "Get PDF Template Placeholders",
        "operationId": "getPDFTemplatePlaceholdersByName",
        "tags": [
          "PDF"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PDFPlaceholders"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find template (template name)"
          }
        },
        "security": []
      }
    },
    "/pdf/templates/{templateName}/fill": {
      "post": {
        "description": "Fill the placeholders of a specific PDF template",
        "summary": "Fill placeholders of PDF template",
        "operationId": "fillPDFTemplate",
        "tags": [
          "PDF"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to fill",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PDFFilledTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name or data supplied"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Placeholder": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        },
        "example": {
          "Name": "Sherlock Holmes",
          "Age": "45"
        }
      },
...
      "EmailSendRequest": {
        "type": "object",
        "properties": {
          "toEmail": {
            "type": "string",
            "example": "holmes@detective.com"
          },
          "toName": {
            "type": "string",
            "example": "Sherlock Holmes"
          },
          "subject": {
            "type": "string",
            "example": "New Case"
          },
          "message": {
            "type": "string",
            "example": "Find the thief!"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "attachmentContent": {
                  "type": "string",
                  "example": "base64 encoded attachment"
                },
                "attachmentName": {
                  "type": "string",
                  "example": "attachment"
                },
                "attachmentExtension": {
                  "type": "string",
                  "example": "txt"
                }
              },
              "required": [
                "attachmentContent",
                "attachmentName",
                "attachmentExtension"
              ]
            }
          }
        },
        "required": [
          "toEmail",
          "toName",
          "subject",
          "message"
        ]
      },
      "EmailTemplate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ExampleName"
          },
          "templateMJMLCode": {
            "type": "string",
            "example": "<mjml><mj-body><mj-section><mj-column><mj-text font-size=\"20px\" color=\"#F45E43\" font-family=\"helvetica\">Hello World</mj-text></mj-column></mj-section></mj-body></mjml>"
          }
        },
        "required": [
          "name",
          "templateMJMLCode"
        ]
      },
      "EmailTemplatePostRequest": {
        "type": "object",
        "properties": {
          "templateString": {
            "type": "string",
            "example": "<mjml><mj-body><mj-section><mj-column><mj-text>Your Name: {{.Name}}, Age: {{.Age}}</mj-text></mj-column></mj-section></mj-body></mjml>"
          },
          "isMJML": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "templateString",
          "isMJML"
        ]
      },
      "EmailTemplateFillSendRequest": {
        "type": "object",
        "properties": {
          "toEmail": {
            "type": "string",
            "example": "holmes@detective.com"
          },
          "toName": {
            "type": "string",
            "example": "Sherlock Holmes"
          },
          "subject": {
            "type": "string",
            "example": "New Case"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "attachmentContent": {
                  "type": "string",
                  "example": "base64 encoded attachment"
                },
                "attachmentName": {
                  "type": "string",
                  "example": "attachment"
                },
                "attachmentExtension": {
                  "type": "string",
                  "example": "txt"
                }
              },
              "required": [
                "attachmentContent",
                "attachmentName",
                "attachmentExtension"
              ]
            }
          },
          "placeholders": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "Name": "Sherlock Holmes",
              "Age": "45"
            }
          }
        },
        "required": [
          "toEmail",
          "toName",
          "subject",
          "placeholders"
        ]
      },
      "EmailPlaceholders": {
        "type": "object",
        "properties": {
          "placeholders": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "key",
                      "value"
                    ],
                    "properties": {
                      "key": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    },
                    "example": [
                      {
                        "key": "firstName",
                        "value": ""
                      },
                      {
                        "key": "lastName",
                        "value": ""
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "required": [
          "placeholders"
        ],
        "example": {
          "placeholders": [
            {
              "data": [
                {
                  "key": "firstName",
                  "value": ""
                },
                {
                  "key": "lastName",
                  "value": ""
                }
              ]
            }
          ]
        }
      },
      "TemplateFillRequest": {
        "type": "object",
        "properties": {
          "placeholders": {
            "$ref": "#/components/schemas/Placeholder"
          }
        },
        "required": [
          "placeholders"
        ]
      },
      "FilledEmailTemplateResponse": {
        "type": "string",
        "example": "<mjml><mj-body><mj-section><mj-column><mj-text font-size=\"20px\" color=\"#F45E43\" font-family=\"helvetica\">Hello World </mj-text></mj-column></mj-section></mj-body></mjml>"
      },
...
      "PDFTemplate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ExampleName"
          },
          "templateString": {
            "type": "string",
            "example": "= Client Name: {{.Name}}, Age: {{.Age}}"
          }
        },
        "required": [
          "name",
          "templateString"
        ]
      },
      "PDFTemplatePostRequest": {
        "type": "object",
        "properties": {
          "templateString": {
            "type": "string",
            "example": "= Client\\nName: {{.Name}}, Age: {{.Age}}"
          }
        },
        "required": [
          "templateString"
        ]
      },
      "PDFPlaceholders": {
        "type": "object",
        "properties": {
          "placeholders": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "key",
                      "value"
                    ],
                    "properties": {
                      "key": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    },
                    "example": [
                      {
                        "key": "firstName",
                        "value": ""
                      },
                      {
                        "key": "lastName",
                        "value": ""
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "required": [
          "placeholders"
        ],
        "example": {
          "placeholders": [
            {
              "data": [
                {
                  "key": "firstName",
                  "value": ""
                },
                {
                  "key": "lastName",
                  "value": ""
                }
              ]
            }
          ]
        }
      },
      "PDFFilledTemplateResponse": {
        "type": "string",
        "example": "base64 encoded PDF file"
      },
...
    }
  }
}